JSGARVEY US276-T278 Send data across group - compress
